+++ FitnessPathApp.Tests/UnitTests/UserServiceTests.cs
@@ -13,7 +13,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace FitnessPathApp.Tests.BusinessLayer
+namespace FitnessPathApp.Tests.UnitTests
 {
     public class UserServiceTests
     {
